+++ test/integration/iaptResultsPage.js
@@ -2,6 +2,7 @@ const chai = require('chai');
 const chaiHttp = require('chai-http');
 const cheerio = require('cheerio');
 
+const config = require('../../config/config');
 const constants = require('../../app/lib/constants');
 const createBody = require('../../app/lib/requests/createBody');
 const getHrefFromA = require('../lib/helpers').getHrefFromA;
@@ -15,7 +16,7 @@ const expect = chai.expect;
 chai.use(chaiHttp);
 
 describe('IAPT results page', () => {
-  const organisationLookupIndex = 'organisationlookup3-index';
+  const organisationLookupIndex = config.api.indexes.orgLookup;
   const path = `/indexes/${organisationLookupIndex}/docs/search`;
   const type = constants.types.IAPT;